Streamlit/requirements.txt
streamlit 
fastapi 
uvicorn
python-dotenv
requests
seaborn
matplotlib
plotly
geopy
toml
boto3
Pillow
st-files-connection
python-jose
pydeck
snowflake-connector-python
numpy
